requirements.txt
fastapi
uvicorn
jinja2
python-dotenv
tortoise-orm
aiohttp
pytz
